+++ AGAT.LocoDispatcher.Parser/Utils/Events/StartMoveEvent.cs
@@ -10,7 +10,7 @@ namespace AGAT.LocoDispatcher.Parser.Utils.Events
         public StartMoveEvent(
             string type,
             int timestamp,
-            int direction,
+            double direction,
             int directionParity,
             string message,
             string trainId,
@@ -18,7 +18,7 @@ namespace AGAT.LocoDispatcher.Parser.Utils.Events
             string checkpointNumber)
         {
             this.Type = type;
-            this.Direction = direction;
+            this.Direction = (int)direction;
             this.DirectionParity = directionParity;
             this.Message = message;
             this.Timestamp = timestamp;
